+++ tests/testthat/test_XcmsExperiment.R
@@ -212,10 +212,19 @@ test_that("saveMsObject,mzTabParam works", {
         saveMsObject(faahko, mzTabParam(studyId = "test_study", path = d,
                                         sampleDataColumn = "sample_name")),
         "has to correspond to column names of the sampleData()")
+    expect_error(saveMsObject(
+        faahko, mzTabParam(studyId = "test_study", path = d,
+                           sampleDataColumn = c("sample_name","sample_index"))),
+        "has to correspond to column names of the sampleData()")
     expect_error(
         saveMsObject(faahko, mzTabParam(studyId = "test_study", path = d,
                                         sampleDataColumn = "sample_index",
                                         optionalFeatureColumns = "other")),
+        "'optionalFeatureColumns' have to correspond")
+    expect_error(saveMsObject(
+        faahko, mzTabParam(studyId = "test_study", path = d,
+                           sampleDataColumn = c("sample_index"),
+                           optionalFeatureColumns = c("mzmed", "mzmin", "a"))),
         "'optionalFeatureColumns' have to correspond")
 
     p <- mzTabParam(studyId = "test_study", path = d,