+++ NEWS.md
@@ -1,5 +1,14 @@
 # Version 0.0
 
+## Changes in 0.0.11
+
+- Fix error message/check for `mzTabParam()` export of `XcmsExperiment`.
+- Use the internal `.retry()` function from *MsBackendMetaboLights* to retry
+  downloading from EBI's ftp server.
+- Avoid duplicated export of `"mzmed"`, `"rtmin"`, `"rtmed"` and `"rtmax"`
+  `featureData()` columns in mzTab-M export.
+- Fix definition of *study_variables* for the SML array in mzTab-M export.
+
 ## Changes in 0.0.10
 
 - Re-use the `Spectra::dataStorageBasePath()` method to update/set the base path